requirements.txt
pandas
requests
tqdm
pillow 
dotenv
opencv-python
plotly
trimesh